+++ CHANGELOG.md
@@ -27,7 +27,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `Pattern.remove_local_clifford_commands` transpiles MBQC+LC patterns into pure MBQC patterns.
   - `Clifford.to_opengraph` returns an open graph without local Cliffords that implements the given single-qubit Clifford gate.
 
-- #515, #519: `Pattern.reindex` returns a pattern whose nodes have been re-indexed either by a supplied mapping or, by default, consecutively starting at 0.
+- #515, #519:
+  - `Pattern.reindex` returns a pattern whose nodes have been re-indexed either by a supplied mapping or, by default, consecutively starting at 0.
+  - `Pattern.node_mapping` returns a mapping that can be passed to `Pattern.reindex`, with flexibility for specifying how nodes are mapped.
 
 ### Fixed